pyproject.toml
[tool.poetry]
name = "ad-creative-performance"
version = "0.1.0"
description = ""
authors = ["motiram <moti@playai.network>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.3"
pytorch-lightning = "^2.4.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
